Add a toml file config wrapper
